I have completed most of the functionality requried in the lab. The only feature I have yet to implement is the voting. All other features are full function (search, view, at to DB). However, I have not included any partials, link_to's, button_to's, and plan to refactor to include those.
For the homework, I have added an actors table and DB, and a joint table. After testing in the terminal, the relationship is funtioning correctly. I just have to create a new movie in the views/ controller and have it save to the DB so I can display actors. Again, I need to refactor to include partials.